src-backend/requirements.txt
requests
pytz
qrcode[pil]
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-filter
psycopg2-binary
celery
flower
redis
django-redis
django-celery-beat
sentry-sdk[django]
sentry-sdk[celery]
